Make excessiveblocksize parameter and setexcessiveblock RPC command conistent (#1747)
* Use uint64_t aux variable in setexcessiveblock for EB

This fix an incositency in the code, in fact the upper limit for EB
via RPC call was 4.29GB (max of uint32_t) whereas it was possible to
set a higher value via bitcoin.conf

* Make sure we use uint64_t for all compuations we do around EB

* Remove useless cast

* Use "Excessive  block size" rather than the EB acronym

[skip ci]

+++ src/qt/unlimiteddialog.cpp
@@ -173,8 +173,8 @@ void UnlimitedDialog::validateBlockSize()
 {
     ui.statusLabel->setStyleSheet("QLabel { color: red; }");
 
-    int mmb = ui.miningMaxBlock->text().toInt();
-    int ebs = excessiveBlockSize;
+    uint64_t mmb = ui.miningMaxBlock->text().toInt();
+    uint64_t ebs = excessiveBlockSize;
 
     if (!MiningAndExcessiveBlockValidatorRule(ebs, mmb))
     {
